CI progress (#445)
* Rearrange devDeps to match Yarn expectations

It appears that the CI functions due to a fluke in how the node modules are resolved. This changes fixes that, providing the devDeps a package(Workspace) needs, rather than relying on it being able to see top-level package.json devDeps.

Add "yarn run ci" to locally run the same thigns as CI will run. Tidy a couple others (explicit "run" along the way.

* Initial E2E app-under-test, very simple

* Test harness for React JS output

Builds; E2E to verify it keeps building is a step forward.

Does not run, because Morpho React JS ESM output does not import React. This can be handled automatically in React JSX bundling, but is expected by downstream needs to be present in JS files.

* Test harness for React TS output

This builds and runs, but requires a ts-ignore. Morpho React TSX output currently expects to be in "source" rather than node_modules, so it doesn't compile the TS(X) and doesn't include .d.ts. This could be enhanced in the future to facilitate easy use of Morpho in one project of a monorepo, with output consume by other projects easily.

* Test harness for React TS - cleanup

* E2E app depends on dependencies needed for targets

Because the morpho output depends on various external libraries (like solid), a morpho project that outputs for those targets must also depend on those libraries.

* Test harness for Solid TS output

Builds, runs.

* e2e-app: export the Nuxt2 Vue3 output as vue3

* Test harness for Vue 3 output

Builds and runs

* Cleanup script names, prettier, e2e-app deps

Since a morpho app output can depend on Vue, a morpho-based project depends on vue.

Prettier.

Unify the script for running an example under the name serve, in preparation for testing them with playwright.

rename preview to serve

prettier

misc

* Remove React-JS E2E; assume TS in 2022.

* Remove "ts" from E2E naming; assume all TS in 2022.

* Recommend playwright and other relevant VSCode extensions

* e2e-app more testable/semantic, ignore test output

* Local working E2E for one target (Solid)

* Run more in CI, including E2E

* E2E for React and Vue also

* Fix CI mistake

* Resolve Vue version warning, reduce a bit of excess logging